requirements.txt
tqdm
opencv-python==4.5.2.52
opencv-python-headless==4.5.2.52
tensorflow
tf-models-official
mediapy
remotezip
seaborn
matplotlib